refactor: Move finish_object function
